src: fix lint on compressed startup blobs
Wrap lines past 80 columns and store the zstd frame size in uint64_t.

+++ src/zstd_blob.cc
@@ -13,7 +13,7 @@ namespace node {
 uint8_t* ZstdDecompressAligned(const uint8_t* src,
                                size_t src_size,
                                size_t* raw_size) {
-  unsigned long long content = ZSTD_getFrameContentSize(src, src_size);
+  uint64_t content = ZSTD_getFrameContentSize(src, src_size);
   if (content == ZSTD_CONTENTSIZE_ERROR ||
       content == ZSTD_CONTENTSIZE_UNKNOWN || content == 0) {
     return nullptr;
